New: move initialisation into Plugin class
